services/lambda_calculator/requirements.txt
aws-lambda-typing==2.20.0
numpy==2.2.3
polars==1.24.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
